AP-2140 Chances of success controller
