+++ app/services/migration_helpers/state_changer.rb
@@ -7,10 +7,10 @@ module MigrationHelpers
       [:client_details_answers_checked, :applicant_details_checked, %(provider_step = 'check_benefits')],
       [:client_details_answers_checked, :provider_entering_means, %|provider_step not in ('check_benefits', 'open_banking_consents', 'email_addresses', 'about_the_financial_assessments')|],
       [:client_details_answers_checked, :provider_confirming_applicant_eligibility, %|provider_step in ('check_benefits', 'open_banking_consents', 'email_addresses', 'about_the_financial_assessments')|],
-      [:provider_assessing_means, :provider_entering_merits, %|provider_step in ('capital_assessment_results', 'start_chances_of_successes', 'date_client_told_incidents', 'respondents', 'statement_of_cases', 'success_likely', 'success_prospects')|],
+      [:provider_assessing_means, :provider_entering_merits, %|provider_step in ('capital_assessment_results', 'start_chances_of_successes', 'date_client_told_incidents', 'respondents', 'statement_of_cases', 'chances_of_success', 'success_prospects')|],
       [:provider_submitted, :applicant_entering_means, nil],
       [:provider_checking_citizens_means_answers, :checking_non_passported_means, nil],
-      [:provider_checked_citizens_means_answers, :provider_entering_merits, %|provider_step not in ('capital_assessment_results', 'start_chances_of_successes', 'date_client_told_incidents', 'respondents', 'statement_of_cases', 'success_likely', 'success_prospects')|]
+      [:provider_checked_citizens_means_answers, :provider_entering_merits, %|provider_step not in ('capital_assessment_results', 'start_chances_of_successes', 'date_client_told_incidents', 'respondents', 'statement_of_cases', 'chances_of_success', 'success_prospects')|]
     ].freeze
 
     def initialize(dummy_run:)
