+++ app/views/providers/proceeding_merits_task/chances_of_success/index.html.erb
@@ -5,7 +5,7 @@
     ) do %>
   <%= form_with(
         model: @form,
-        url: providers_legal_aid_application_success_likely_index_path,
+        url: providers_legal_aid_application_chances_of_success_index_path,
         method: :post,
         local: true
       ) do |form| %>
