+++ app/views/providers/success_prospects/show.html.erb
@@ -1,7 +1,7 @@
 <%= page_template(
       page_title: t('.h1-heading'),
       template: :basic,
-      back_link: { path: providers_legal_aid_application_success_likely_index_path(@legal_aid_application) }
+      back_link: { path: providers_legal_aid_application_chances_of_success_index_path(@legal_aid_application) }
     ) do %>
   <%= render(
         'shared/forms/success_prospect_form',
