+++ app/views/shared/check_answers/_merits.html.erb
@@ -130,7 +130,7 @@
   <dl class="govuk-summary-list govuk-!-margin-bottom-2">
     <%= check_answer_link(
           name: :success_likely,
-          url: providers_legal_aid_application_success_likely_index_path(@legal_aid_application),
+          url: providers_legal_aid_application_chances_of_success_index_path(@legal_aid_application),
           question: t('.items.prospects_of_success'),
           answer: yes_no(chances_of_success.success_likely),
           read_only: read_only
@@ -140,7 +140,7 @@
   <dl class="govuk-summary-list govuk-!-margin-bottom-2">
     <%= check_answer_link(
           name: :success_prospect,
-          url: providers_legal_aid_application_success_likely_index_path(@legal_aid_application),
+          url: providers_legal_aid_application_chances_of_success_index_path(@legal_aid_application),
           question: t('.items.success_prospect'),
           answer: t("shared.forms.success_prospect.#{chances_of_success.success_prospect}"),
           read_only: read_only
